+++ src/besselLogLikelihood.cpp
@@ -260,60 +260,3 @@ double BesselLogLikelihood::EvaluateLFunction(const double sqDist, const double 
   resVal *= boost::math::cyl_bessel_j((double)m_DomainDimension / 2.0, tmpVal);
   return resVal;
 }
-
-double BesselLogLikelihood::GetLogDeterminant()
-{
-  arma::mat lMatrix(m_SampleSize, m_SampleSize);
-  arma::mat lMatrixDeriv1(m_SampleSize, m_SampleSize);
-  arma::mat lMatrixDeriv2(m_SampleSize, m_SampleSize);
-  arma::mat lMatrixDeriv3(m_SampleSize, m_SampleSize);
-  arma::mat lMatrixDeriv4(m_SampleSize, m_SampleSize);
-  double resVal = 0.0;
-  double workValue1 = 0.0;
-  double workValue2 = 0.0;
-  double workValue3 = 0.0;
-  double workValue4 = 0.0;
-  double workSign = 0.0;
-
-  for (unsigned int i = 0;i < m_SampleSize;++i)
-  {
-    for (unsigned int j = i;j < m_SampleSize;++j)
-    {
-      double sqDist = m_DistanceMatrix(i, j) * m_DistanceMatrix(i, j);
-      unsigned int workLabel = m_PointLabels[i] + m_PointLabels[j];
-
-      if (workLabel == 2)
-        resVal = this->EvaluateLFunction(sqDist, m_FirstIntensity, m_FirstAmplitude, m_FirstAlpha);
-      else if (workLabel == 3)
-        resVal = this->EvaluateLFunction(sqDist, m_CrossIntensity, m_CrossAmplitude, m_CrossAlpha);
-      else
-        resVal = this->EvaluateLFunction(sqDist, m_SecondIntensity, m_SecondAmplitude, m_SecondAlpha);
-
-      lMatrix(i, j) = resVal;
-      lMatrixDeriv1(i, j) = workValue1;
-      lMatrixDeriv2(i, j) = workValue2;
-      lMatrixDeriv3(i, j) = workValue3;
-      lMatrixDeriv4(i, j) = workValue4;
-
-      if (i != j)
-      {
-        lMatrix(j, i) = resVal;
-        lMatrixDeriv1(j, i) = workValue1;
-        lMatrixDeriv2(j, i) = workValue2;
-        lMatrixDeriv3(j, i) = workValue3;
-        lMatrixDeriv4(j, i) = workValue4;
-      }
-    }
-  }
-
-  arma::log_det(resVal, workSign, lMatrix);
-  arma::mat lMatrixInverse = arma::inv(lMatrix);
-
-  m_GradientLogDeterminant.set_size(this->GetNumberOfParameters());
-  m_GradientLogDeterminant[0] = arma::trace(lMatrixInverse * lMatrixDeriv1);
-  m_GradientLogDeterminant[1] = arma::trace(lMatrixInverse * lMatrixDeriv2);
-  m_GradientLogDeterminant[2] = arma::trace(lMatrixInverse * lMatrixDeriv3);
-  m_GradientLogDeterminant[3] = arma::trace(lMatrixInverse * lMatrixDeriv4);
-
-  return resVal;
-}
