+++ src/besselLogLikelihood.h
@@ -5,24 +5,7 @@
 class BesselLogLikelihood : public BaseLogLikelihood
 {
 public:
-  BesselLogLikelihood() : BaseLogLikelihood()
-  {
-    m_FirstAlpha = 1.0;
-    m_SecondAlpha = 1.0;
-    m_CrossAlpha = 1.0;
-    m_FirstIntensity = 0.0;
-    m_SecondIntensity = 0.0;
-    m_CrossIntensity = 0.0;
-    m_FirstAmplitude = 0.0;
-    m_SecondAmplitude = 0.0;
-    m_CrossAmplitude = 0.0;
-    m_EstimateFirstBValue = true;
-    m_EstimateSecondBValue = true;
-    m_EstimateCrossBValue = true;
-    m_EstimateFirstBetaValue = true;
-    m_EstimateSecondBetaValue = true;
-    m_EstimateCrossBetaValue = true;
-  }
+  BesselLogLikelihood() : BaseLogLikelihood() {}
 
   void SetFirstAlpha(const double x);
   void SetSecondAlpha(const double x);
@@ -37,11 +20,4 @@ private:
   bool CheckModelParameters(const arma::mat &params);
   double EvaluateLFunction(const double sqDist, const double intensity, const double amplitude, const double alpha);
   double GetIntegral();
-  double GetLogDeterminant();
-
-  double m_FirstAlpha, m_CrossAlpha, m_SecondAlpha;
-  double m_FirstIntensity, m_CrossIntensity, m_SecondIntensity;
-  double m_FirstAmplitude, m_CrossAmplitude, m_SecondAmplitude;
-  bool m_EstimateFirstBValue, m_EstimateCrossBValue, m_EstimateSecondBValue;
-  bool m_EstimateFirstBetaValue, m_EstimateCrossBetaValue, m_EstimateSecondBetaValue;
 };
